riscv: fix build errors in RVV patches for ffmpeg-spacemit
- configure: register spacemit_mpp in EXTERNAL_LIBRARY_LIST; add help text; disable vulkan (broken glslc)
- libswscale/riscv/swscale.c: fix SwsContext->SwsInternal type mismatch; add missing nv12/nv21 forward decls
- libavfilter/riscv/vf_nlmeans_init_riscv.c: fix int->ptrdiff_t mismatch in weights_line declaration
- libavfilter/vf_overlay.c: fix ARCH_RISCV init call: use main->format and !overlay_straight

+++ libavfilter/riscv/vf_nlmeans_init_riscv.c
@@ -7,6 +7,7 @@
  */
 
 #include "config.h"
+#include <stddef.h>
 #include "libavutil/attributes.h"
 #include "libavutil/riscv/cpu.h"
 #include "libavfilter/vf_nlmeans.h"
@@ -25,8 +26,8 @@ void ff_compute_weights_line_rvv(const uint32_t *const iia,
                                   float *total_weight,
                                   float *sum,
                                   const float *const weight_lut,
-                                  int max_meaningful_diff,
-                                  int startx, int endx);
+                                  ptrdiff_t max_meaningful_diff,
+                                  ptrdiff_t startx, ptrdiff_t endx);
 
 av_cold void ff_nlmeans_init_riscv(NLMeansDSPContext *dsp)
 {
